requirements.txt
# Core dependencies for RAG project
streamlit>=1.25.0
faiss-cpu>=1.7.4
ollama>=0.1.7

# Deep Learning & Transformers
# IMPORTANT: transformers 4.45.0 is pinned for compatibility with PyTorch 2.5.1
# If using PyTorch >= 2.6.0, you can upgrade to transformers >= 4.56.0
transformers>=4.45.0,<4.56.0
torch>=2.5.0,<2.6.0
torchvision>=0.20.0
torchaudio>=2.5.0
sentence-transformers>=2.2.2

# HuggingFace ecosystem
huggingface-hub>=0.16.4
datasets>=2.14.0
tokenizers>=0.13.3
safetensors>=0.4.1

# Document processing and text manipulation
langchain>=0.0.300
langchain-community>=0.0.30
pypdf2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21

# Utilities and data handling
numpy>=1.24.0,<2.4.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
tqdm>=4.65.0
rank-bm25>=0.2.2

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Optional: For enhanced functionality
chromadb>=0.4.0
pinecone-client>=2.2.2
weaviate-client>=3.22.0

# Development tools (optional)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Image processing
pillow>=9.0.0
